feature(validate-signup): validate user input
This commit adds the validations required to make the test cases pass. The vaidation checks for invalid email, checks if email or username already exists, password is not alphanumeric, password is less than 8 characters, username is not empty.
A middleware has been created which checks the user input.
The model is also been validated.

Files changed/modified:
- package.json
- .editorconfig
- server/models/userModels.test.js
- server/models/user.js
- server/test/user.test.js
- server/controllers/userController.js
- server/routes/api/user.js

Start #160451498