feature(rate-articles): write rating fucntionality
This commit rates an article. When a valid number or float is sent to the rate api endpoint, an article will be rated and its average computed. This average consists of all the ratings for an article and it is stored int the articles table.
A test to check if the rating sent is a valid number or float has also been added.

Files changed:
- server/tests/article.test.js
- server/tests/rate.test.js
- server/controller./rateController.js
- server/models/article.js
- server/migrations/20181003095441-create-article.js

Finishes #160451503