feature(tag-articles): tag and untag articles
This commit adds the tagging and untagging of articles.

Notable changes include:
Tags has been added to the articles request. So when an article is created or updated, the tags sent along with it are associated with it.
A new helper file has been created for joi validation to extract some common functionality.
More test cases have been added.

Finishes #160451508